requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.12.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
